[VIEW] package -> writing javadocs for each method
