+++ src/main/java/view/ViewManager.java
@@ -1,12 +1,12 @@
 package view;
 
-import interface_adapter.ViewManagerModel;
-
 import java.awt.CardLayout;
 import java.beans.PropertyChangeEvent;
 import java.beans.PropertyChangeListener;
 
 import javax.swing.JPanel;
+
+import interface_adapter.ViewManagerModel;
 
 /**
  * The View Manager for the program. It listens for property change events
@@ -17,6 +17,12 @@ public class ViewManager implements PropertyChangeListener {
     private final JPanel views;
     private final ViewManagerModel viewManagerModel;
 
+    /**
+     * Constructs a ViewManager that manages switching between views.
+     * @param views the panel containing all views
+     * @param cardLayout the CardLayout used to switch views
+     * @param viewManagerModel the model that notifies which view should be shown
+     */
     public ViewManager(JPanel views, CardLayout cardLayout, ViewManagerModel viewManagerModel) {
         this.views = views;
         this.cardLayout = cardLayout;
@@ -24,6 +30,10 @@ public class ViewManager implements PropertyChangeListener {
         this.viewManagerModel.addPropertyChangeListener(this);
     }
 
+    /**
+     * Handles property change events to update the visible view when the state changes.
+     * @param evt the property change event
+     */
     @Override
     public void propertyChange(PropertyChangeEvent evt) {
         if (evt.getPropertyName().equals("state")) {
